[Refact] #26 DataInitializer에 point 추가

+++ src/main/java/com/umust/dobonglife/domain/user/domain/entity/User.java
@@ -51,5 +51,5 @@ public class User extends BaseEntity {
     private String providerId;
 
     @Column
-    private int balance;
+    private long balance;
 }
